backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain-openai==0.3.29
langchain-core==0.3.74
langgraph==0.6.4
pdfplumber==0.11.7
pydantic==2.11.7
openai==1.99.6